refactor: reading 0 오류 해결 배럴 파일에서 유틸을 가져올 경우 해당 문제가 생김 추후 이유를 정확히 파악해 볼 예정
